+++ tgbotapi.behaviour_builder.fsm/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/BehaviourContextWithFSM.kt
@@ -9,6 +9,8 @@ import dev.inmo.micro_utils.coroutines.accumulatorFlow
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.*
+import kotlin.jvm.JvmName
+import kotlin.reflect.KClass
 
 /**
  * Interface which combine [BehaviourContext] and [StatesMachine]. Subcontext of triggers and states contexts must have
@@ -26,9 +28,30 @@ interface BehaviourContextWithFSM<T : State> : BehaviourContext, StatesMachine<T
         handlers: List<BehaviourWithFSMStateHandlerHolder<*, T>>
     ): T? {
         return handlers.firstOrNull { it.checkHandleable(state) } ?.run {
-            handleState(contextUpdatesFlow, state)
+            doInSubContext(updatesUpstreamFlow = contextUpdatesFlow) {
+                handleState(state)
+            }
         }
     }
+
+    /**
+     * Add NON STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Non strict means that
+     * for input [State] will be used [KClass.isInstance] and any inheritor of [kClass] will pass this requirement
+     *
+     * @see BehaviourWithFSMStateHandlerHolder
+     * @see onStateOrSubstate
+     */
+    fun <I : T> add(kClass: KClass<I>, strict: Boolean = false, handler: BehaviourWithFSMStateHandler<I, T>)
+
+    /**
+     * Add STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Strict means that
+     * for input [State] will be used [State]::class == [kClass] and any [State] with exactly the same type will pass
+     * requirements
+     *
+     * @see BehaviourWithFSMStateHandlerHolder
+     * @see strictlyOn
+     */
+    fun <I : T> addStrict(kClass: KClass<I>, handler: BehaviourWithFSMStateHandler<I, T>) = add(kClass, strict = true, handler)
 
     override fun copy(
         bot: TelegramBot,
@@ -48,6 +71,28 @@ interface BehaviourContextWithFSM<T : State> : BehaviourContext, StatesMachine<T
     }
 }
 
+
+/**
+ * Add NON STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Non strict means that
+ * for input [State] will be used [KClass.isInstance] and any inheritor of [kClass] will pass this requirement
+ *
+ * @see BehaviourWithFSMStateHandlerHolder
+ * @see BehaviourContextWithFSM.add
+ */
+@Suppress("MemberVisibilityCanBePrivate")
+inline fun <reified I : O, O: State> BehaviourContextWithFSM<O>.onStateOrSubstate(handler: BehaviourWithFSMStateHandler<I, O>) = add(I::class, strict = false, handler)
+
+/**
+ * Add STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Strict means that
+ * for input [State] will be used [State]::class == [kClass] and any [State] with exactly the same type will pass
+ * requirements
+ *
+ * @see BehaviourWithFSMStateHandlerHolder
+ * @see BehaviourContextWithFSM.addStrict
+ */
+@Suppress("MemberVisibilityCanBePrivate")
+inline fun <reified I : O, O: State> BehaviourContextWithFSM<O>.strictlyOn(handler: BehaviourWithFSMStateHandler<I, O>) = addStrict(I::class, handler)
+
 /**
  * Default realization of [BehaviourContextWithFSM]. It uses [behaviourContext] as a base for this object as
  * [BehaviourContext], but managing substates contexts updates for avoiding of updates lost between states
@@ -58,6 +103,9 @@ class DefaultBehaviourContextWithFSM<T : State>(
     private val handlers: List<BehaviourWithFSMStateHandlerHolder<*, T>>
 ) : BehaviourContext by behaviourContext, BehaviourContextWithFSM<T> {
     private val updatesFlows = mutableMapOf<Any, Flow<Update>>()
+    private val additionalHandlers = mutableListOf<BehaviourWithFSMStateHandlerHolder<*, T>>()
+    private var actualHandlersList = additionalHandlers + handlers
+
     private fun getContextUpdatesFlow(context: Any) = updatesFlows.getOrPut(context) {
         allUpdatesFlow.accumulatorFlow(scope)
     }
@@ -65,12 +113,17 @@ class DefaultBehaviourContextWithFSM<T : State>(
     override suspend fun StatesMachine<in T>.handleState(state: T): T? = launchStateHandling(
         state,
         allUpdatesFlow,
-        handlers
+        actualHandlersList
     )
+
+    override fun <I : T> add(kClass: KClass<I>, strict: Boolean, handler: BehaviourWithFSMStateHandler<I, T>) {
+        additionalHandlers.add(BehaviourWithFSMStateHandlerHolder(kClass, strict, handler))
+        actualHandlersList = additionalHandlers + handlers
+    }
 
     override fun start(scope: CoroutineScope): Job = scope.launchSafelyWithoutExceptions {
         val statePerformer: suspend (T) -> Unit = { state: T ->
-            val newState = launchStateHandling(state, getContextUpdatesFlow(state.context), handlers)
+            val newState = launchStateHandling(state, getContextUpdatesFlow(state.context), actualHandlersList)
             if (newState != null) {
                 statesManager.update(state, newState)
             } else {
@@ -94,6 +147,26 @@ class DefaultBehaviourContextWithFSM<T : State>(
             launch { statePerformer(it) }
         }
     }
+    /**
+     * Add NON STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Non strict means that
+     * for input [State] will be used [KClass.isInstance] and any inheritor of [kClass] will pass this requirement
+     *
+     * @see BehaviourWithFSMStateHandlerHolder
+     * @see BehaviourContextWithFSM.add
+     */
+    @Suppress("MemberVisibilityCanBePrivate")
+    inline fun <reified I : T> onStateOrSubstate(handler: BehaviourWithFSMStateHandler<I, T>) = add(I::class, strict = false, handler)
+
+    /**
+     * Add STRICT [handler] to list of available in future [BehaviourContextWithFSM]. Strict means that
+     * for input [State] will be used [State]::class == [kClass] and any [State] with exactly the same type will pass
+     * requirements
+     *
+     * @see BehaviourWithFSMStateHandlerHolder
+     * @see BehaviourContextWithFSM.addStrict
+     */
+    @Suppress("MemberVisibilityCanBePrivate")
+    inline fun <reified I : T> strictlyOn(handler: BehaviourWithFSMStateHandler<I, T>) = addStrict(I::class, handler)
 
     override suspend fun startChain(state: T) {
         statesManager.startChain(state)
@@ -106,7 +179,7 @@ class DefaultBehaviourContextWithFSM<T : State>(
         onBufferOverflow: BufferOverflow,
         upstreamUpdatesFlow: Flow<Update>?,
         updatesFilter: BehaviourContextAndTypeReceiver<Boolean, Update>?
-    ): BehaviourContextWithFSM<T> = BehaviourContextWithFSM(
+    ): DefaultBehaviourContextWithFSM<T> = BehaviourContextWithFSM(
         behaviourContext.copy(bot, scope, broadcastChannelsSize, onBufferOverflow, upstreamUpdatesFlow, updatesFilter),
         handlers,
         statesManager